requirements.txt
typer==0.12.3
rich==13.7.1
pyyaml==6.0.2
tldextract==5.1.2
beautifulsoup4==4.12.3
lxml==5.3.0
jinja2==3.1.4
pydantic==2.8.2
tabulate==0.9.0

# Gmail / OAuth
google-api-python-client==2.142.0
google-auth==2.34.0
google-auth-oauthlib==1.2.1

# Optional server
fastapi==0.115.0
uvicorn==0.30.6

# Dev / test
pytest==8.3.2
pytest-cov==5.0.0
